moved all playlist components under feed
