+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/feed/history-view/history-view.component.scss
