+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/feed/history-view/history-view.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { VideoCardComponent } from '../video-displays/video-card/video-card.component';
+import { VideoCardComponent } from '../../video-displays/video-card/video-card.component';
 import { NgFor } from '@angular/common';
 
 @Component({
