+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/feed/playlist/playlist-mini/playlist-mini.component.scss
