+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/feed/playlist/playlist-view/playlist-view.component.html
