+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/module/playlist/playlist-routing.module.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { PlaylistViewComponent } from '../../components/playlist/playlist-view/playlist-view.component';
+import { PlaylistViewComponent } from '../../components/feed/playlist/playlist-view/playlist-view.component';
 
 export const playlistRoutes: Routes = [
   {
